+++ src/test/java/org/rere/examples/parameterMatching/ParameterMatchingExampleExpected.java
@@ -17,7 +17,7 @@ public class ParameterMatchingExampleExpected {
 
   public static ParameterMatchingExample.PrivateDice environmentNode0() throws Exception {
     ParameterMatchingExample.PrivateDice mockObject = mock(ParameterMatchingExample.PrivateDice.class);
-    doReturn(10).when(mockObject).roll(anyInt(), anyDouble(), anyLong(), anyShort(), anyChar(), anyByte(), anyBoolean(), anyFloat(), anyString());
+    doReturn((int) 10).when(mockObject).roll(anyInt(), anyDouble(), anyLong(), anyShort(), anyChar(), anyByte(), anyBoolean(), anyFloat(), anyString());
     return mockObject;
   }
 }
